rename _omnisci to omnisci_backend
