+++ rbc/omnisci_backend/numpy_funcs.py
