+++ rbc/tests/test_omnisci_array_methods.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-from rbc._omnisci import Array
+from rbc.omnisci_backend import Array
 
 
 rbc_omnisci = pytest.importorskip('rbc.omniscidb')
